Magana/requirements.txt
flask
requests
python-dotenv
google-genai
gradio_client
PyPDF2
python-docx
